# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RateMaster API",
    "description": "Revenue/pricing recommendation SaaS for hotels",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/data/import/preview": {
      "post": {
        "tags": [
          "data"
        ],
        "summary": "Import Preview",
        "description": "Preview CSV headers for column mapping. Returns headers and auto-detected mapping.",
        "operationId": "import_preview_api_v1_data_import_preview_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_preview_api_v1_data_import_preview_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Import Preview Api V1 Data Import Preview Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/data/import": {
      "post": {
        "tags": [
          "data"
        ],
        "summary": "Import Csv",
        "description": "Upload CSV with column mapping. Creates snapshot + rows.",
        "operationId": "import_csv_api_v1_data_import_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_csv_api_v1_data_import_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataSnapshotResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/data/snapshots": {
      "get": {
        "tags": [
          "data"
        ],
        "summary": "List Snapshots",
        "description": "List data snapshots for user's properties.",
        "operationId": "list_snapshots_api_v1_data_snapshots_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "property_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Property Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DataSnapshotResponse"
                  },
                  "title": "Response List Snapshots Api V1 Data Snapshots Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_import_csv_api_v1_data_import_post": {
        "properties": {
          "property_id": {
            "type": "string",
            "title": "Property Id"
          },
          "snapshot_type": {
            "type": "string",
            "title": "Snapshot Type",
            "default": "current"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "column_mapping": {
            "type": "string",
            "title": "Column Mapping",
            "default": ""
          },
          "snapshot_date": {
            "type": "string",
            "title": "Snapshot Date",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "property_id",
          "file"
        ],
        "title": "Body_import_csv_api_v1_data_import_post"
      },
...
      "Body_import_preview_api_v1_data_import_preview_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_preview_api_v1_data_import_preview_post"
      },
      "DataSnapshotResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "property_id": {
            "type": "string",
            "title": "Property Id"
          },
          "snapshot_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Date"
          },
          "snapshot_type": {
            "type": "string",
            "title": "Snapshot Type"
          },
          "row_count": {
            "type": "integer",
            "title": "Row Count"
          },
          "data_health_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Health Score"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "recommended_fixes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommended Fixes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "property_id",
          "snapshot_date",
          "snapshot_type",
          "row_count",
          "data_health_score",
          "created_at"
        ],
        "title": "DataSnapshotResponse",
        "description": "Data snapshot response."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}